deprecation: moves get_vectors_mesh over to vector_utils
