+++ pyxem/utils/virtual_images_utils.py
@@ -39,7 +39,7 @@ def normalize_virtual_images(im):
     return imn
 
 
-@deprecated(since="0.18.0", removal="1.0.0")
+@deprecated(since="0.18.0", removal="1.0.0",alternative="utils.vectors.get_vector_mesh")
 def get_vectors_mesh(g1_norm, g2_norm, g_norm_max, angle=0.0, shear=0.0):
     """
     Calculate vectors coordinates of a mesh defined by a norm, a rotation and
@@ -63,31 +63,5 @@ def get_vectors_mesh(g1_norm, g2_norm, g_norm_max, angle=0.0, shear=0.0):
         x and y coordinates of the vectors of the mesh
 
     """
-
-    def rotation_matrix(angle):
-        return np.array(
-            [[np.cos(angle), -np.sin(angle)], [np.sin(angle), np.cos(angle)]]
-        )
-
-    def shear_matrix(shear):
-        return np.array([[1.0, shear], [0.0, 1.0]])
-
-    if shear < 0 or shear > 1:
-        raise ValueError("The `shear` value must be in the interval [0, 1].")
-
-    order1 = int(np.ceil(g_norm_max / g1_norm))
-    order2 = int(np.ceil(g_norm_max / g2_norm))
-    order = max(order1, order2)
-
-    x = np.arange(-g1_norm * order, g1_norm * (order + 1), g1_norm)
-    y = np.arange(-g2_norm * order, g2_norm * (order + 1), g2_norm)
-
-    xx, yy = np.meshgrid(x, y)
-    vectors = np.stack(np.meshgrid(x, y)).reshape((2, (2 * order + 1) ** 2))
-
-    transformation = rotation_matrix(np.radians(angle)) @ shear_matrix(shear)
-
-    vectors = transformation @ vectors
-    norm = np.linalg.norm(vectors, axis=0)
-
-    return vectors[:, norm <= g_norm_max].T
+    from vector_utils import get_vectors_mesh
+    return get_vectors_mesh(g1_norm,g2_norm,g_norm_max,angle,shear)
